@@ -1,6 +1,6 @@
 {
   "name": "i18n-po-json",
-  "version": "1.0.6",
+  "version": "1.0.7",
   "description": "i18n .po to .json file converter",
   "main": "dist/index.js",
   "scripts": {
@@ -22,27 +22,29 @@
   "license": "MIT",
   "devDependencies": {
     "@types/get-stdin": "^5.0.1",
+    "@types/mocha": "^5.2.5",
+    "@types/node": "8.9.4",
     "@types/yargs": "^8.0.2",
     "array-xor": "^0.1.1",
-    "browserify": "^14.4.0",
+    "browserify": "^16.2.2",
     "eslint": "3.13.1",
     "i18n-proto": "1.0.5",
-    "karma": "1.7.0",
+    "karma": "^3.0.0",
     "karma-browserify": "^5.1.1",
     "karma-firefox-launcher": "^1.0.1",
     "karma-mocha": "^1.3.0",
-    "karma-typescript": "^3.0.4",
-    "mocha": "^3.4.2",
-    "ts-node": "^3.3.0",
-    "tslib": "1.5.0",
+    "karma-typescript": "^3.0.13",
+    "mocha": "^5.2.0",
+    "ts-node": "^7.0.1",
+    "tslib": "^1.9.3",
     "tslint": "4.3.1",
     "tslint-eslint-rules": "3.2.3",
-    "typescript": "^2.4.2",
-    "uglify-js": "3.0.18",
+    "typescript": "^2.9.2",
+    "uglify-js": "^3.4.9",
     "watchify": "^3.9.0"
   },
   "dependencies": {
-    "get-stdin": "5.0",
-    "yargs": "8.0"
+    "get-stdin": "6.0.0",
+    "yargs": "8.0.0"
   }
 }
